+++ packages/keyring-snap-bridge/src/v2/index.ts
@@ -1,1 +1,2 @@
 export * from './SnapKeyring';
+export * from './SnapKeyringV1Adapter';